Fixed bug where we were accidentally masking error
In lib/fake_ifdh, we were masking a bug where if you say "return '' and int('')>0",
for example, we'd expect to get a ValueError from the second conditional, but instead
an empty string is returned.  I fixed the function so that we explicitly only do the
second check, and reraise the ValueError.

Also fixed a bug where we don't try to read the tokenfile if it doesn't exist, since
os.popen(anything_with_nonexistent_file) will simply return an empty string if we save
its results into a file-like object.