Make --verbose and --debug backward-compatible with old jobsub_client
Normally, we would have chosen to make a --verbose flag, and have different --debug flag levels, but since the legacy jobsub_client had only a --debug flag, we chose to make jobsub_lite backward-compatible in that way, and have a --debug flag, with --verbose having the various levels.  --debug is equivalent to --verbose 1.