Sanitize error message for MAX_PROCS_PER_SUBMISSION; keep schedd ads in memory during command execution
If a user tries to submit with too many procs (the -N flag), then we need to
tell them clearly (a) That they're trying to submit too many jobs, and (b) What
the limit actually is.

To do this, in the old way, we'd have to query the collector twice during a submission.
This commit implements a global variable that stores the schedd ads, so that rather than
query the collector each time, we can look locally if the info is available.