break jobsub_submit main into two functions
The first (main()) does basic arg parsing, and creates an overarching tracing span for the submission.
The second (submit()) does the rest.

The primary motivation is to create a container tracing span for the submit steps,
so they do not clutter the trace graph and provide a clear separation between submit
and execution stages.